+++ src/commands/deploy/shouldValidateConnection.ts
@@ -4,20 +4,17 @@ import { ConnectionKey, DurableBackend, DurableBackendValues } from "../../const
 import { getEventHubName } from "../appSettings/connectionSettings/eventHubs/validateEventHubsConnection";
 
 export interface IShouldValidateConnection {
-    shouldValidateStorage: boolean;
     shouldValidateEventHubs: boolean;
     shouldValidateSqlDb: boolean;
 }
 
 export async function shouldValidateConnections(durableStorageType: DurableBackendValues | undefined, client: SiteClient, projectPath: string): Promise<IShouldValidateConnection> {
     const app: StringDictionary = await client.listApplicationSettings();
-    const remoteStorageConnection: string | undefined = app?.properties?.[ConnectionKey.Storage];
     const remoteEventHubsConnection: string | undefined = app?.properties?.[ConnectionKey.EventHubs];
     const remoteSqlDbConnection: string | undefined = app?.properties?.[ConnectionKey.SQL];
     const eventHubName: string | undefined = await getEventHubName(projectPath);
 
-    const shouldValidateStorage: boolean = !remoteStorageConnection;
     const shouldValidateEventHubs: boolean = durableStorageType === DurableBackend.Netherite && (!eventHubName || !remoteEventHubsConnection);
     const shouldValidateSqlDb: boolean = durableStorageType === DurableBackend.SQL && !remoteSqlDbConnection;
-    return { shouldValidateStorage, shouldValidateEventHubs, shouldValidateSqlDb };
+    return { shouldValidateEventHubs, shouldValidateSqlDb };
 }